ml-pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pyyaml>=6.0
boto3>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0
joblib>=1.3.0
pytest>=7.4.0
